requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
dash==0.26.5
dash-core-components==0.28.0
dash-html-components==0.12.0
dash-renderer==0.13.2
decorator==4.3.0
enum34==1.1.6
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Compress==1.4.0
functools32==3.2.3.post2
gunicorn==19.9.0
idna==2.7
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.15.1
pandas==0.23.4
plotly==3.2.1
psycopg2==2.7.5
pyaml==17.12.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
retrying==1.3.3
six==1.11.0
traitlets==4.3.2
urllib3==1.23
Werkzeug==0.14.1
